refactor: move RobotManifest from roz-copper to roz-core
RobotManifest is pure data + serde parsing with no IO or Copper
dependencies. Belongs in roz-core alongside ChannelManifest.
roz-copper re-exports for backward compatibility.
