+++ crates/roz-cli/Cargo.toml
@@ -59,7 +59,6 @@ tonic-build.workspace = true
 
 [dev-dependencies]
 tempfile = "3"
-roz-copper = { path = "../roz-copper", default-features = false }
 
 [lints]
 workspace = true
